+++ docs/database/add_contacts_email_constraint.sql
@@ -32,3 +32,12 @@ COMMENT ON INDEX contacts_franchise_email_unique_idx IS 'Unique constraint for u
 
 
 
+
+
+
+
+
+
+
+
+