+++ docs/database/add_county_to_territory_checks.sql
@@ -18,3 +18,12 @@ COMMENT ON COLUMN territory_checks.county IS 'County name for territory check lo
 
 
 
+
+
+
+
+
+
+
+
+