+++ docs/database/add_llm_parsing_fields.sql
@@ -32,3 +32,12 @@ COMMENT ON COLUMN scraping_runs.llm_parsing_completed_at IS 'Timestamp when LLM 
 
 
 
+
+
+
+
+
+
+
+
+