+++ docs/database/add_llm_processed_at.sql
@@ -25,3 +25,12 @@ COMMENT ON COLUMN franchises.llm_processed_at IS 'Timestamp when franchise was p
 
 
 
+
+
+
+
+
+
+
+
+