+++ frontend/src/types/react-simple-maps.d.ts
@@ -60,3 +60,12 @@ declare module 'react-simple-maps' {
 
 
 
+
+
+
+
+
+
+
+
+